requirements.txt
smbprotocol==1.10.1
watchdog==3.0.0
pyyaml==6.0.1
requests==2.31.0
pillow==10.0.0
click==8.1.7
